feat: merge the `sender` and `severity` in the DDS payload into the same field to simplify the protocol

+++ yazi-dds/src/body/body.rs
@@ -76,7 +76,7 @@ impl<'a> Body<'a> {
 	}
 
 	#[inline]
-	pub fn with_severity(self, severity: u8) -> Payload<'a> {
+	pub fn with_severity(self, severity: u16) -> Payload<'a> {
 		Payload::new(self).with_severity(severity)
 	}
 }
